# openapiOidc.json
{
  "swagger": "2.0",
  "info": {
    "title": "greet",
    "version": "1.0.0",
    "description": "description"
  },
  "paths": {
    "/hello": {
      "get": {
        "tags": [
          "hello"
        ],
        "summary": "Get Hello",
        "description": "Get Simple Hello",
        "operationId": "GetHello",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Hello"
            }
          }
        }
      }
    },
    "/hello/{name}": {
      "get": {
        "tags": [
          "hello"
        ],
        "summary": "Show hello with a given name",
        "description": "Show hello with a given name",
        "operationId": "GetHelloName",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string",
              "items": {
                "$ref": "#/definitions/Hello"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Given Name",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/goodbye": {
      "get": {
        "tags": [
          "goodbye"
        ],
        "summary": "Get Goodbye",
        "description": "Get Simple Goodbye",
        "operationId": "GetGoodbye",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Goodbye"
            }
          }
        }
      }
    },
    "/goodbye/{name}": {
      "get": {
        "tags": [
          "goodbye"
        ],
        "summary": "Show goodbye with a given name",
        "description": "Show goodbye with a given name",
        "operationId": "GetGoodbyeName",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string",
              "items": {
                "$ref": "#/definitions/Goodbye"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Given Name",
            "required": true,
            "type": "string"
          }
        ]
      }
    }
  },
  "components": {
    "oauth2": {
      "security_auth": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "http://sso-rh-sso.apps.cluster-8fhm9.8fhm9.sandbox1428.opentlc.com/auth/realms/user11/protocol/openid-connect/auth",
            "tokenUrl": "http://sso-rh-sso.apps.cluster-8fhm9.8fhm9.sandbox1428.opentlc.com/auth/realms/user11/protocol/openid-connect/token",
            "refreshUrl": "http://dev-games-challenge-5-user11.apps.cluster-8fhm9.8fhm9.sandbox1428.opentlc.com"
          }
        }
      }
    }
  },
  "security": [
    {
      "oauth2": [
      ]
    }
  ]
}